% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_chunk_utils.R
\name{is_reactive_call}
\alias{is_reactive_call}
\alias{is_reactive_val_call}
\alias{is_reactive_values_call}
\alias{is_any_reactive_call}
\alias{is_variable_call}
\alias{is_input_call}
\alias{is_session_user_data}
\title{Call Checks}
\usage{
is_reactive_call(x, env = rlang::caller_env())

is_reactive_val_call(x, env = rlang::caller_env())

is_reactive_values_call(x, env = rlang::caller_env())

is_any_reactive_call(x, env = rlang::caller_env())

is_variable_call(x, existing_vars = NULL, env = rlang::caller_env())

is_input_call(x)

is_session_user_data(x)
}
\arguments{
\item{x}{An R call object}

\item{env}{The environment the call is being made, by default it is the environment
calling the check, but is likely the environment the call is being made i.e. the
reactive expression.}

\item{existing_vars}{A character vector of variable definitions that exist in the \code{Repro} object}
}
\value{
A boolean value determining whether or not the call check has passed.
}
\description{
A set of helper functions that determine what type of call is being made within
an expression.

\code{is_reactive_call} checks whether or not the call is evaluating a
\code{shiny::reactive} variable.

\code{is_reactive_val_call} checks whether or not the call is evaluating a
\code{shiny::reactiveVal} variable.

\code{is_reactive_values_call} checks whether or not the call is evaluating an item
within a \code{shiny::reactiveValues} variable.

\code{is_any_reactive_call} checks whether or not the call points to evaluating a
\code{reactive}, \code{reactiveVal} or \code{reactiveValues}.

\code{is_variable_call} checks whether or not the call point to a variable that is defined
within the given module.

\code{is_input_call} checks whether or not the call points to evaluate an input value.

\code{is_session_user_data} checks whether or not the call points to evaluate an object
within \code{session$userData}
}
\keyword{internal}
