% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repro_S7.R
\name{is_new_reactive}
\alias{is_new_reactive}
\title{Reactive Variables Definition Check}
\usage{
is_new_reactive(new, exisitng)
}
\arguments{
\item{new, exisitng}{A named list of reactive variable definitions}
}
\value{
A boolean stating whether or not there is at least one reactive definition
in \code{new} that doesn't exist in \code{existing}
}
\description{
A helper function to check whether or not the reactive variables to be added
to the \code{Repro} object already exists. Used to avoid duplicate definitions being
added to a script.
}
\keyword{internal}
