% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repro_chunk.R
\name{repro_chunk}
\alias{repro_chunk}
\title{Reproduce Code Chunk}
\usage{
repro_chunk(x, repro_code = Repro(), env = rlang::caller_env())
}
\arguments{
\item{x}{\code{\link[shiny:reactive]{shiny::reactive()}} object to make reproducible}

\item{repro_code}{A \code{Repro} object to store calls found in \code{x}. By default it is
empty, but if \code{x} is not the first call within an expression, this will have
prior calls and pre-requisites that might be used in \code{x}.}

\item{env}{The environment \code{x} is defined in. By default it is the environment of where
\code{reprex_reactive} is called}
}
\value{
A \code{Repro} object containing all the necessary code and packages to recreate
the provided expression when evaluated.
}
\description{
Evaluate a chunk of code to extract Shiny inputs and reactives, replacing the inputs with
the values selected by the user, and the reactives with the code bodies used to generate
them.
}
\details{
Whilst a default is provided to \code{env}, it is unlikely that this is the same environment
\code{x} is defined in. This allows the top-level \code{reprex_reactive} call to pass through
environments found for calls to other reactives in the chunk.
}
\keyword{internal}
