% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Repro_S7.R
\name{Repro}
\alias{Repro}
\title{Reproducible Code}
\usage{
Repro(code = list(), packages = character(0), prerequisites = list())
}
\arguments{
\item{code}{Code chunks found in a given expression}

\item{packages}{Packages found in the function calls in the code and/or pre-requisites}

\item{prerequisites}{Code chunks used to generate reactive objects found in the code}
}
\description{
An S7 object that holds the code and packages required to re-create a given reactive.
}
\keyword{internal}
